+++ mud-contracts/world-v2/src/systems/smart-character/ISmartCharacterErrors.sol
@@ -1,6 +1,0 @@
-//SPDX-License-Identifier: MIT
-pragma solidity >=0.8.24;
-
-interface ISmartCharacterErrors {
-  error SmartCharacter_ERC721AlreadyInitialized();
-}