+++ mud-contracts/world-v2/test/inventory/InventoryTest.t.sol
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+pragma solidity >=0.8.24;
+
+import "forge-std/Test.sol";
+import { MudTest } from "@latticexyz/world/test/MudTest.t.sol";
+
+contract InventoryTest is MudTest {}